# Conflicts:
#	Changelog
#	RELEASE
#	RELEASE_NOTES
#	configure
#	doc/Doxyfile
#	libavcodec/cbs_av1.c
#	libavcodec/cbs_av1_syntax_template.c
#	libavcodec/h264_redundant_pps_bsf.c
#	libavcodec/imm4.c
#	libavcodec/jvdec.c
#	libavcodec/mjpegbdec.c
#	libavcodec/prosumer.c
#	libavfilter/Makefile
#	libavfilter/version.h
#	libavfilter/vf_tonemap_opencl.c
#	libavfilter/vf_yadif_cuda.c
#	libavfilter/yadif_common.c
#	libavformat/aacdec.c
#	libavformat/matroskaenc.c
#	tests/api/api-h264-slice-test.c

+++ libavcodec/ilbcdec.c
@@ -724,7 +724,7 @@ static void construct_vector (
     int16_t cbvec0[SUBL];
     int16_t cbvec1[SUBL];
     int16_t cbvec2[SUBL];
-    int32_t a32;
+    unsigned a32;
     int16_t *gainPtr;
     int j;
 
@@ -747,7 +747,7 @@ static void construct_vector (
         a32 += SPL_MUL_16_16(*gainPtr++, cbvec1[j]);
         a32 += (unsigned)SPL_MUL_16_16(*gainPtr, cbvec2[j]);
         gainPtr -= 2;
-        decvector[j] = (a32 + 8192) >> 14;
+        decvector[j] = (int)(a32 + 8192) >> 14;
     }
 }
